Added super command MigrateAltinn2FormSchemas
