--- RepoCleanup/Application/Commands/CommitChangesCommand.cs
@@ -1,12 +1,0 @@
-﻿namespace RepoCleanup.Application.Commands
-{
-    public class CommitChangesCommand
-    {
-        public CommitChangesCommand(string localPath)
-        {
-            LocalPath = localPath;
-        }
-
-        public string LocalPath { get; private set; }
-    }
-}
